Add Tomcat6Connector

+++ src/main/scala/cc/spray/Settings.scala
@@ -20,6 +20,6 @@ import akka.config.Config._
 
 object Settings {
   lazy val RootActorId         = config.getString("spray.root-actor-id", "spray-root-service")
-  lazy val AsyncTimeout        = config.getLong("spray.timeout", 1000)
+  lazy val AsyncTimeout        = config.getInt("spray.timeout", 1000)
   lazy val CloseConnection     = config.getBool("spray.close-connection", true)
 }
